Removed gradient in quadtree
The idea behind the gradient was to make a quick check if data varies too much before doing a
PCA for the area. However, this was done solely on the z-axis, which can vary while still
being flat in x- or y-direction. It is therefore removed, and parameters has been tuned to this
change. As a side note, this speeds up the algorithm with about 2ms.